backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
email-validator==2.2.0
firebase-admin==6.5.0
google-cloud-aiplatform==1.71.1
google-cloud-dlp==3.22.0
google-cloud-firestore==2.19.0
google-cloud-kms==3.1.0
pinecone-client==5.0.1
elasticsearch==8.15.1
aiohttp==3.10.10
tiktoken==0.8.0
httpx==0.27.2
python-multipart==0.0.12
pypdf==5.0.1
python-docx==1.1.2
pytest==8.3.3
pytest-asyncio==0.24.0
